+++ memory_op.h
@@ -1,7 +1,7 @@
 #ifndef MEMORY_OP_H
 #define MEMORY_OP_H
 
-void load(short& accumulator, short* main_memory, short mem_addr);
-void store(short& accumulator, short* main_memory, short mem_addr);
+void load(int& accumulator, int* main_memory, int mem_addr);
+void store(int& accumulator, int* main_memory, int mem_addr);
 
 #endif
